api/backup/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9

# AI/LLM
openai==1.3.7

# Data processing
numpy==1.24.3
langdetect==1.0.9

# HTTP and file handling
python-multipart==0.0.6
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Environment and configuration
python-dotenv==1.0.0
